Improved delAll* fplugins

Merged-on: https://assembla.com/code/faraday/git/merge_requests/4633283

+++ bin/delAllHost.py
@@ -14,6 +14,14 @@ __prettyname__ = 'Delete All Hosts'
 
 
 def main(workspace='', args=None, parser=None):
+    parser.add_argument('-y', '--yes', action="store_true")
+    parsed_args = parser.parse_args(args)
+    if not parsed_args.yes:
+        msg = ("Are you sure you want to delete all hosts in the "
+               "workspace {}? This action can't be undone [y/n] ".format(
+                   workspace))
+        if raw_input(msg) not in ('y', 'yes'):
+            return 1, None
     for host in models.get_hosts(workspace):
         print('Delete Host:' + host.name)
         models.delete_host(workspace, host.id)
